Export package added
New Package Added which will export the data, when created, from MSSQL to a local MariaDB or MySQL database.

Cannot yet establish an ODBC connection to a remote MariaDB database. This is for a future release

Note that the SCRIPTS/MARIADB folder now contains an SQL script that can be run on any MariaDB or MySQL database to create the tables.

I haven't been able to automate this so that it runs from Visual Studio (ODBC driver gives an error '.COM not registered' and reports a syntax error where none exiss)